+++ app/src/main/java/com/example/buggyweather/home/repository/CurrentWeatherDataSource.kt
@@ -1,6 +1,6 @@
 package com.example.buggyweather.home.repository
 
-import com.example.buggyweather.domain.CurrentWeather
+import com.example.buggyweather.core.domain.CurrentWeather
 
 interface CurrentWeatherDataSource {
 	suspend fun getCurrentWeather(cityName: String, unitsName: String): CurrentWeather
