+++ app/src/main/java/com/example/buggyweather/home/usecase/GetCurrentWeatherUseCase.kt
@@ -1,8 +1,8 @@
 package com.example.buggyweather.home.usecase
 
 import com.example.buggyweather.core.base.UseCase
-import com.example.buggyweather.domain.CurrentWeather
-import com.example.buggyweather.domain.MeasuringUnits
+import com.example.buggyweather.core.domain.CurrentWeather
+import com.example.buggyweather.core.domain.MeasuringUnits
 import com.example.buggyweather.home.repository.CurrentWeatherDataSource
 
 class GetCurrentWeatherUseCase(
