+++ app/src/main/java/com/example/buggyweather/main/repository/MeasuringUnitsRepository.kt
@@ -1,7 +1,7 @@
 package com.example.buggyweather.main.repository
 
 import android.content.SharedPreferences
-import com.example.buggyweather.domain.MeasuringUnits
+import com.example.buggyweather.core.domain.MeasuringUnits
 
 class MeasuringUnitsRepository(
 		private val settingSharedPref: SharedPreferences
