+++ app/src/main/java/com/example/buggyweather/whole/presenter/WholeDayViewModel.kt
@@ -5,10 +5,10 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
 import com.example.buggyweather.core.base.BaseViewModel
 import com.example.buggyweather.core.base.UseCase
-import com.example.buggyweather.domain.Coordinate
-import com.example.buggyweather.domain.ForecastRequest
-import com.example.buggyweather.domain.MeasuringUnits
-import com.example.buggyweather.domain.WholeDayWeather
+import com.example.buggyweather.core.domain.Coordinate
+import com.example.buggyweather.core.domain.ForecastRequest
+import com.example.buggyweather.core.domain.MeasuringUnits
+import com.example.buggyweather.core.domain.WholeDayWeather
 import com.example.buggyweather.core.network.exception.RemoteException
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
