+++ app/src/main/java/com/example/buggyweather/whole/presenter/adapter/ForecastListViewHolder.kt
@@ -2,8 +2,8 @@ package com.example.buggyweather.whole.presenter.adapter
 
 import androidx.recyclerview.widget.RecyclerView
 import com.example.buggyweather.databinding.ItemForecastBinding
-import com.example.buggyweather.domain.HourlyWeather
-import com.example.buggyweather.domain.MeasuringUnits
+import com.example.buggyweather.core.domain.HourlyWeather
+import com.example.buggyweather.core.domain.MeasuringUnits
 import com.example.buggyweather.core.utils.epochToDateTime
 import com.example.buggyweather.core.utils.loadWeatherIcon
 import kotlin.math.roundToInt
