+++ app/src/main/java/com/example/buggyweather/whole/repository/WholeDayWeatherDataSource.kt
@@ -1,7 +1,7 @@
 package com.example.buggyweather.whole.repository
 
-import com.example.buggyweather.domain.ForecastRequest
-import com.example.buggyweather.domain.WholeDayWeather
+import com.example.buggyweather.core.domain.ForecastRequest
+import com.example.buggyweather.core.domain.WholeDayWeather
 
 interface WholeDayWeatherDataSource {
 	suspend fun getWholeDayWeather(request: ForecastRequest): WholeDayWeather
