+++ app/src/test/java/com/example/buggyweather/home/presenter/HomeViewModelTest.kt
@@ -1,8 +1,8 @@
 package com.example.buggyweather.home.presenter
 
 import com.example.buggyweather.core.base.UseCase
-import com.example.buggyweather.domain.CurrentWeather
-import com.example.buggyweather.domain.MeasuringUnits
+import com.example.buggyweather.core.domain.CurrentWeather
+import com.example.buggyweather.core.domain.MeasuringUnits
 import com.example.buggyweather.helper.BaseTest
 import com.example.buggyweather.helper.getOrAwaitValue
 import com.example.buggyweather.core.network.KnownExceptionMessage
