+++ app/src/test/java/com/example/buggyweather/home/usecase/GetCurrentWeatherUseCaseTest.kt
@@ -1,8 +1,8 @@
 package com.example.buggyweather.home.usecase
 
 import com.example.buggyweather.core.base.UseCase
-import com.example.buggyweather.domain.CurrentWeather
-import com.example.buggyweather.domain.MeasuringUnits
+import com.example.buggyweather.core.domain.CurrentWeather
+import com.example.buggyweather.core.domain.MeasuringUnits
 import com.example.buggyweather.helper.BaseTest
 import com.example.buggyweather.home.repository.CurrentWeatherDataSource
 import com.example.buggyweather.core.network.KnownExceptionMessage
