+++ app/src/test/java/com/example/buggyweather/main/repository/MeasuringUnitsRepositoryTest.kt
@@ -1,7 +1,7 @@
 package com.example.buggyweather.main.repository
 
 import android.content.SharedPreferences
-import com.example.buggyweather.domain.MeasuringUnits
+import com.example.buggyweather.core.domain.MeasuringUnits
 import com.example.buggyweather.helper.BaseTest
 import com.nhaarman.mockitokotlin2.doReturn
 import com.nhaarman.mockitokotlin2.verify
